Update EmailSenderExtensions.cs
Could be my ignorance, but this code did not work for me when the link had parameters separated by ampersands. This proposed change solved the problem for me, and then I went (curious) to see if I could find where this code snippet had come from. And here I am... thought I'd share my findings. Apologies if I'm mistaken and this proposed change is not an improvement.